pyproject.toml
[tool.poetry]
name = "news_provider"
version = "0.1.0"
description = ""
authors = ["Kyzard0 <diego_rafael_97@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Django = "^3.2.4"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.8.6"
Pillow = "^8.2.0"
django-rest-auth = "^0.9.5"
django-allauth = "^0.44.0"
djangorestframework-simplejwt = "^4.7.1"

[tool.poetry.dev-dependencies]
ipython = "^7.24.1"
django-debug-toolbar = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
